+++ tests/geojson/test_feature_collection.py
@@ -1,0 +1,61 @@
+import typing
+import pytest
+
+from pydantic import Field
+from pydantic_shapely import FeatureBaseModel, GeometryField
+from pydantic_shapely.geojson.feature_collection import GeoJsonFeatureCollectionBaseModel
+from shapely.geometry import Point
+
+
+class FeatureModel(FeatureBaseModel, geometry_field="point"):
+    point: typing.Annotated[Point, GeometryField(), Field(...)]
+    name: str = "Hello World"
+    answer: int = 42
+
+
+expected_result = """{
+  "type": "FeatureCollection",
+  "features": [
+    {
+      "type": "Feature",
+      "geometry": {
+        "type": "Point",
+        "coordinates": [
+          0.0,
+          0.0
+        ]
+      },
+      "properties": {
+        "name": "Hello World",
+        "answer": 42
+      }
+    },
+    {
+      "type": "Feature",
+      "geometry": {
+        "type": "Point",
+        "coordinates": [
+          1.0,
+          1.0
+        ]
+      },
+      "properties": {
+        "name": "Hello World",
+        "answer": 42
+      }
+    }
+  ]
+}"""
+
+def test_create_feature_collection_roundtrip():
+    test = GeoJsonFeatureCollectionBaseModel[FeatureModel.GeoJsonDataModel].from_feature_models(
+        [
+            FeatureModel(point=Point(0, 0)),
+            FeatureModel(point=Point(1, 1)),
+        ]
+    )
+    assert test.model_dump_json(indent=2) == expected_result
+    assert test.to_feature_models() == [
+        FeatureModel(point=Point(0, 0)),
+        FeatureModel(point=Point(1, 1)),
+    ]